requirements.txt
# Core LLM
openai>=1.0.0

# LangGraph & LangChain
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Vector DB
chromadb>=0.5.0

# Token counting (bonus: better than word count)
tiktoken>=0.7.0

# Environment
python-dotenv>=1.0.0
